fix accidental column change

+++ database/migrations/2025_05_25_035536_remove_nullable_from_shift_fields.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::table('shifts', function (Blueprint $table) {
             $table->float('multiplier')->change();
             $table->float('length')->change();
-            $table->float('num_spots')->change();
+            $table->unsignedInteger('num_spots')->change();
         });
     }
 
@@ -26,7 +26,7 @@ return new class extends Migration
         Schema::table('shifts', function (Blueprint $table) {
             $table->float('multiplier')->nullable()->change();
             $table->float('length')->nullable()->change();
-            $table->float('num_spots')->nullable()->change();
+            $table->unsignedInteger('num_spots')->nullable()->change();
         });
     }
 };
